requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-libtorrent>=2.0.9
python-jose[cryptography]==3.3.0
bcrypt>=4.0
python-multipart==0.0.9
slowapi==0.1.9
pydantic>=2.0,<3
pydantic-settings>=2.0
httpx==0.27.0
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
# For PostgreSQL: asyncpg>=0.29
